rename file, add telemetry event to status bar menu open
