+++ packages/core/src/codewhisperer/activation.ts
@@ -60,7 +60,7 @@ import { notifyNewCustomizations } from './util/customizationUtil'
 import { CodeWhispererCommandBackend, CodeWhispererCommandDeclarations } from './commands/gettingStartedPageCommands'
 import { SecurityIssueHoverProvider } from './service/securityIssueHoverProvider'
 import { SecurityIssueCodeActionProvider } from './service/securityIssueCodeActionProvider'
-import { listCodeWhispererCommands } from './ui/statusBar'
+import { listCodeWhispererCommands } from './ui/statusBarMenu'
 import { updateUserProxyUrl } from './client/agent'
 import { Container } from './service/serviceContainer'
 const performance = globalThis.performance ?? require('perf_hooks').performance
